Final Project/requirements.txt
altair
ansi2html
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
comm
contourpy
cycler
dash
dash-core-components
dash-html-components
dash-table
debugpy
decorator
defusedxml
exceptiongroup
executing
fastjsonschema
Flask
fonttools
fqdn
gitdb
GitPython
h11
httpcore
httpx
idna
importlib-metadata
ipykernel
ipython
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
jmespath
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-dash
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
nbclient
nbconvert
nbformat
nest-asyncio
nltk
notebook_shim
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
Pillow
platformdirs
plotly
prometheus_client
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pycparser
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
python-json-logger
pytz
PyYAML
pyzmq
referencing
regex
requests
retrying
rfc3339-validator
rfc3986-validator
rich
rpds-py
s3transfer
scikit-learn
scipy
Send2Trash
six
smmap
sniffio
soupsieve
stack-data
streamlit
tenacity
terminado
threadpoolctl
tinycss2
toml
tomli
toolz
tornado
traitlets
types-python-dateutil
typing_extensions
tzdata
tzlocal
uri-template
urllib3
vaderSentiment
validators
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
zipp